refactor: separate tasks files in task folder
