+++ src/domain/usecases/task/add-task-usecase.js
@@ -1,5 +1,3 @@
-const { MissingParamError } = require("../../utils/errors")
-
 module.exports = class AddTaskUseCase {
   constructor({ addTaskRepository } = {}) {
     this.addTaskRepository = addTaskRepository
