+++ src/domain/usecases/task/load-task-by-id-usecase.spec.js
@@ -1,5 +1,3 @@
-const { MissingParamError } = require("../../utils/errors")
-const { TaskNotFoundError } = require("../errors")
 const LoadTaskByIdUseCase = require("./load-task-by-id-usecase")
 
 const makeFakeTaskResult = () => ({
