+++ src/infra/repositories/task/load-tasks-by-user-id-db-repository.js
@@ -1,4 +1,4 @@
-const TaskDbModel = require("../database/sequelize/models/task-database-model")
+const TaskDbModel = require("../../database/sequelize/models/task-database-model")
 
 module.exports = class LoadTasksByUserIdDbRepository { 
   async loadAllTask(user_id) {
