+++ src/main/factory/controllers/task/add-task-controller-impl.js
@@ -1,6 +1,6 @@
-const AddTaskUseCase = require("../../../domain/usecases/add-task-usecase")
-const AddTaskDbRepository = require("../../../infra/repositories/add-task-db-repository")
-const AddTaskController = require("../../../presentation/controllers/add-task-controller")
+const AddTaskUseCase = require("../../../../domain/usecases/task/add-task-usecase")
+const AddTaskDbRepository = require("../../../../infra/repositories/task/add-task-db-repository")
+const AddTaskController = require("../../../../presentation/controllers/task/add-task-controller")
 
 const makeAddTaskController = () => {
   const addTaskDbRepository = new AddTaskDbRepository()
