+++ src/main/factory/controllers/task/load-task-by-id-controller-impl.js
@@ -1,6 +1,6 @@
-const LoadTaskByIdController = require("../../../presentation/controllers/load-task-by-id-controller")
-const LoadTaskByIdUseCase = require('../../../domain/usecases/load-task-by-id-usecase')
-const LoadTaskByIdDbRepository = require("../../../infra/repositories/load-task-by-id-db-repository")
+const LoadTaskByIdController = require("../../../../presentation/controllers/task/load-task-by-id-controller")
+const LoadTaskByIdUseCase = require('../../../../domain/usecases/task/load-task-by-id-usecase')
+const LoadTaskByIdDbRepository = require("../../../../infra/repositories/task/load-task-by-id-db-repository")
 
 const makeLoadTaskByIdController = () => {
   const loadTaskByIdDbRepository = new LoadTaskByIdDbRepository()
