+++ src/main/factory/controllers/task/load-tasks-by-user-id-controller-impl.js
@@ -1,6 +1,6 @@
-const LoadTasksByUserIdUseCase = require('../../../domain/usecases/load-tasks-by-user-id-usecase')
-const LoadTasksByUserIdDbRepository = require('../../../infra/repositories/load-tasks-by-user-id-db-repository')
-const LoadTasksByUserIdController = require("../../../presentation/controllers/load-tasks-by-user-id-controller")
+const LoadTasksByUserIdUseCase = require('../../../../domain/usecases/task/load-tasks-by-user-id-usecase')
+const LoadTasksByUserIdDbRepository = require('../../../../infra/repositories/task/load-tasks-by-user-id-db-repository')
+const LoadTasksByUserIdController = require("../../../../presentation/controllers/task/load-tasks-by-user-id-controller")
 
 const makeLoadTasksByUserIdController = () => {
   const loadTasksByUserIdDbRepository = new LoadTasksByUserIdDbRepository()
